TYRUS-393: Incorrect handling of sending whole message during (unfinished) partial message.
Specification allows implementation to thrown an IllegalStateException, but it is not requires - meaning that implementation can even block the thread and wait until the message is fully sent. Current change combines these two approaches - we will wait (up to defined timeout - currently 3 s) and if the message is not finished, the send* call will fail with IllegalStateException.